Proper cleanup
This sends INT to the e2e tests so they can cleanup properly.
e2e.test does not respond well to TERM so we capture TERM in
the parent process and send an INT to the child instead.

Also disable log dumping. We have a sonobuoy plugin that does that.

Related to heptio/sonobuoy#214

Signed-off-by: Chuck Ha <ha.chuck@gmail.com>